refactor ginko module
